Refactor: KeywordItem 으로 쓰던 컴포넌트에서 Recent/Suggestion으로 분리
