+++ src/components/search/RecentKeywordItem/RecentKeywordItem.styled.js
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+const RecentKeyword = styled.div`
+  display: flex;
+  align-items: center;
+  vertical-align: middle;
+  height: 25px;
+  margin: 10px 0px;
+  :hover {
+    background-color: azure;
+  }
+  column-gap: 10px;
+`;
+
+export { RecentKeyword };
